requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
alembic
pydantic-settings
pytest
httpx
asyncpg
strawberry-graphql[fastapi]
grpcio
grpcio-tools